gallery page - details view
Pull events into the Backbone event handler section.

Make render dependent on a custom event instead of a promise so we don't need to create promises in every function that wants to fetch the child details.

Fix the navigation, and make it work off the current order of models in the Backbone collection to allow for ordering the collection and navigating to the correct next/prev child.  This was done by checking the next element in the DOM in the past which didn't work...  At all.

Clean up pretty much every other function in the view.  It's almost all better now.