creates Outside Contact Group model
This model is meant to be a container for outside contacts.  Any arbitrary group name can be given, and contacts added.  It was handled this way because mailing lists were acting as group containers in the past, and they should really be just for mailing lists.  These groups can now be added to mailing lists to mass include outside contacts in e-mail blasts.