rewrite ALL Express views and rework all locals handling for templating
This was a MASSIVE rewrite of views to remove async and replace it with native Promises, it moves assignment on locals, and processing of returned data, from the services into the Express views where it belongs.  The end result is faster load times across the site and much more reusability of service functions, as they’re no longer making assumptions as to how to data is meant to be used, and instead returning it as is and allowing the calling functions to modify and use it.

NOTE: This only removes async in the views, the same will be done to model and other non-view related services soon.