#249 ( WIP ) | create child change history import function
Where I left off:

The first entry in each set of changes has name and other information that may have changed.  I’ll need to determine how to save those entries.  First thought is load the child, and check the fields that always appearsuch as name, if it hasn't changed, don't update it.  You'll need to load the child, create a copy of the child make changes to the copy, then pass both into the change history function with the date and admin to associate it with.

For consistency, I’ll need to save the first update chronologically as ‘record created’ meaning don’t pass in the original Child model.  If this isn’t the case, I’ll need to update the child ( and family ) imports to save the first change history entry if it’s not already happening.