fix list display showing incorrect dates (#179)
This was causing dates to be stored at the specified date with a time of 00:00:00.000Z in UTC, which was interpreted as -4:00:00 when displayed in date columns.

Note: in order to correct the issue for previously saved dates, 5 hours need to be added to each date field in the database.