Fixes memory leak in child matching middleware
The first call to the child matching middleware requires empty criteria because when any search criteria is set the middleware will try to find matching families instead of returning the child data to set search defaults.  The fix ensures no criteria can be specified when only a child id is passed.