Adds defensive code around child gallery data parsing
This code was failing silently due to a child without a gender value in the database, which prevented the child gallery from displaying.  This should not be possible as gender is a required field, but the try/catch block will allow the child gallery to still render if this situation somehow occurs again in the future.